pyproject.toml
[project]
name = "aoc-25"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "advent-of-code-data>=2.1.0",
    "python-dotenv>=1.2.1",
]
